feat(api): Add user authentication and registration endpoints
The AuthController now includes endpoints for user registration and
login. This allows users to create new accounts and authenticate
themselves using their username and password.

The registration process includes validation of the provided data to
ensure that the username is available and that the password meets
the minimum requirements.

The login process authenticates users against the database and
issues JWT tokens upon successful authentication. These tokens can be
used to access protected resources throughout the application.

+++ src/main/java/com/example/hotelmanagement/controller/assembler/RoomAssembler.java
@@ -1,7 +1,9 @@
 package com.example.hotelmanagement.controller.assembler;
 
+import com.example.hotelmanagement.constants.ApplicationRole;
 import com.example.hotelmanagement.controller.RoomController;
 import com.example.hotelmanagement.dto.response.ResponseBase;
+import com.example.hotelmanagement.helper.StaticHelper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
@@ -9,6 +11,7 @@ import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -16,15 +19,23 @@ import java.util.stream.StreamSupport;
 @RequiredArgsConstructor
 public class RoomAssembler {
     public <T extends ResponseBase> EntityModel<T> toRoomModel(T responseDto, Authentication authentication) {
-        return EntityModel.of(responseDto,
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).getRoomById(responseDto.getId(), null)).withSelfRel().withType("GET"),
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).getRoomAllRooms(null)).withRel("Get all room").withType("GET"),
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).createRoom(null, authentication)).withRel("Create new room").withType("POST"),
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).updateRoom(responseDto.getId(), null, authentication)).withRel("Update room").withType("PUT"),
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).deleteRoom(responseDto.getId())).withRel("Delete room").withType("DELETE")
-        );
-    }
+        Set<String> roles = StaticHelper.extractAllRoles(authentication);
 
+        EntityModel<T> entityModel = EntityModel.of(responseDto,
+                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).getRoomById(responseDto.getId(), null)).withSelfRel().withType("GET"),
+                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).getRoomAllRooms(null)).withRel("Get all room").withType("GET")
+                );
+
+        if (roles.contains(ApplicationRole.ADMIN.name()) || roles.contains(ApplicationRole.MANAGER.name())) {
+            entityModel.add(
+                    WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).createRoom(null, authentication)).withRel("Create new room").withType("POST"),
+                    WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).updateRoom(responseDto.getId(), null, authentication)).withRel("Update room").withType("PUT"),
+                    WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomController.class).deleteRoom(responseDto.getId())).withRel("Delete room").withType("DELETE")
+            );
+        }
+
+        return entityModel;
+    }
 
     public <T extends ResponseBase> CollectionModel<EntityModel<T>> toCollectionModel(Iterable<T> responseDtos, Authentication authentication) {
         return StreamSupport.stream(responseDtos.spliterator(), false) //
