+++ src/main/java/com/example/hotelmanagement/controller/assembler/RoomTypeAssembler.java
@@ -1,26 +1,36 @@
 package com.example.hotelmanagement.controller.assembler;
 
+import com.example.hotelmanagement.constants.ApplicationRole;
 import com.example.hotelmanagement.controller.RoomTypeController;
 import com.example.hotelmanagement.dto.response.ResponseBase;
+import com.example.hotelmanagement.helper.StaticHelper;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
 @Component
 public class RoomTypeAssembler {
     public <T extends ResponseBase> EntityModel<T> toModel(T entity, Authentication authentication) {
-        return EntityModel.of(entity,
+        Set<String> roles = StaticHelper.extractAllRoles(authentication);
+        EntityModel<T> entityModel = EntityModel.of(entity,
                 WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).getRoomTypeById(entity.getId(), authentication)).withSelfRel().withType("GET"),
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).getRoomAllTypes(authentication)).withRel("Get all room types").withType("GET"),
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).createRoomType(null, authentication)).withRel("Create new room type").withType("POST"),
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).updateRoomType(entity.getId(), null, authentication)).withRel("Update room type").withType("PUT"),
-                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).deleteRoomType(entity.getId())).withRel("Delete room type").withType("DELETE")
+                WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).getRoomAllTypes(authentication)).withRel("Get all room types").withType("GET")
         );
+        if (roles.contains(ApplicationRole.ADMIN.name()) || roles.contains(ApplicationRole.MANAGER.name())) {
+            entityModel.add(
+                    WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).createRoomType(null, authentication)).withRel("Create new room type").withType("POST"),
+                    WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).updateRoomType(entity.getId(), null, authentication)).withRel("Update room type").withType("PUT"),
+                    WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(RoomTypeController.class).deleteRoomType(entity.getId())).withRel("Delete room type").withType("DELETE")
+            );
+        }
+
+        return entityModel;
     }
 
     public <T extends ResponseBase> CollectionModel<EntityModel<T>> toCollectionModel(Iterable<T> entities, Authentication authentication) {
